+++ packages/akashic-cli-lib-manage/src/update/update.ts
@@ -34,7 +34,7 @@ export function _completeUpdateParameterObject(param: UpdateParameterObject): vo
 
 export function promiseUpdate(param: UpdateParameterObject): Promise<void> {
 	_completeUpdateParameterObject(param);
-	var restoreDirectory = cmn.Util.chdir(param.cwd);
+	const restoreDirectory = cmn.Util.chdir(param.cwd);
 	return Promise.resolve()
 		.then(() => param.debugNpm.update(param.moduleNames))
 		.then(() => param.logger.info("Done!"))